initial cleanup of App.js file
